Pareto view file organization updates
