+++ web/packages/studio/src/components/charts/ExperimentGroupParetoChart/utils.ts
